chore: move form-type component
