+++ src/components/form-type/configured-form-type.jsx
@@ -1,7 +1,7 @@
 import { reduxForm } from 'redux-form';
 
 import FormType from './form-type.component';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'type',
