+++ src/components/form-type/field-checkbox-type/field-checkbox.component.tsx
